Backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
click==8.3.0
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
